+++ cli/commands/handler/delete.go
@@ -4,13 +4,14 @@ import (
 	"fmt"
 
 	"github.com/sensu/sensu-go/cli"
+	"github.com/sensu/sensu-go/cli/commands/helpers"
 	"github.com/sensu/sensu-go/types"
 	"github.com/spf13/cobra"
 )
 
 // DeleteCommand adds a command that allows user to delete handlers
 func DeleteCommand(cli *cli.SensuCli) *cobra.Command {
-	return &cobra.Command{
+	cmd := &cobra.Command{
 		Use:          "delete [NAME]",
 		Short:        "delete handlers given name",
 		SilenceUsage: true,
@@ -22,6 +23,13 @@ func DeleteCommand(cli *cli.SensuCli) *cobra.Command {
 			}
 
 			name := args[0]
+			if skipConfirm, _ := cmd.Flags().GetBool("skip-confirm"); !skipConfirm {
+				if confirmed := helpers.ConfirmDelete(name, cmd.OutOrStdout()); !confirmed {
+					fmt.Fprintln(cmd.OutOrStdout(), "Canceled")
+					return nil
+				}
+			}
+
 			handler := &types.Handler{Name: name}
 			err := cli.Client.DeleteHandler(handler)
 			if err != nil {
@@ -32,4 +40,8 @@ func DeleteCommand(cli *cli.SensuCli) *cobra.Command {
 			return nil
 		},
 	}
+
+	cmd.Flags().Bool("skip-confirm", false, "skip interactive confirmation prompt")
+
+	return cmd
 }
